Remove all calls to SDL's "render" subsystem.
SDL provides an abstraction for 2D graphics, but it's quirky and
incomplete (No way to draw rotated images or set a clipping rectangle,
weird heavy-weight textures...). Every platform we'll be deploying on has
OpenGL anyways, so let's just continue using that and enjoy the benefits
of more available features later.