Ported Amity to webOS!
The test app runs perfectly on my Pre 2, as do the Flambe demos. After
abstracting away the Android specific stuff, the stage is now set for
iOS...

webOS includes a version of SDL 1.2 as part of its PDK. Amity for
Android uses 1.3 (as will iOS), but it was fairly easy to branch in
places for 1.2. I believe you can even use the 1.2 API in SDL 1.3 and
things will just work, but I'd rather embrace the future. Once the PDK
gets 1.3, that compatibility can be removed.